third_party/WebKit/LayoutTests/inspector/sources/debugger/show-generator-location.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/console-test.js"></script>
<script>

function forward(iter, step)
{
    while (step-- > 0)
        iter.next();
    return iter;
}

function* gen()
{
    yield 1;
    yield 2;
    yield 3;
}

var iterNotStarted = gen();
var iterSuspended1 = forward(gen(), 1);
var iterSuspended2 = forward(gen(), 2);
var iterSuspended3 = forward(gen(), 3);
var iterClosed = forward(gen(), 4);

var old = testRunner.closeWebInspector;
testRunner.closeWebInspector = function()
{
    testRunner.closeWebInspector = old;
    testRunner.closeWebInspector();
    testRunner.logToStderr("closeWebInspector");
}

function test()
{
    InspectorTest.startDumpingProtocolMessages();
    var panel = WebInspector.panels.sources;

    function performStandardTestCase(pageExpression, next)
    {
        InspectorTest.addSniffer(panel, "showUISourceCode", showUISourceCodeHook);
        InspectorTest.evaluateInPage(pageExpression, didEvaluate);

        function didEvaluate(remote)
        {
            panel._showGeneratorLocation(remote);
        }

        function showUISourceCodeHook(uiSourceCode, lineNumber, columnNumber, forceShowInPanel)
        {
            // lineNumber and columnNumber are 0-based
            ++lineNumber;
            ++columnNumber;
            InspectorTest.addResult("Generator location revealed: [" + lineNumber + ":" + columnNumber + "]");
            next();
        }
    }

    var expressions = [
        "iterNotStarted",
        "iterSuspended1",
        "iterSuspended2",
        "iterSuspended3",
        "iterClosed",
    ];

    function createTestSuiteFunction(expression)
    {
        var functionName = "test" + expression.toTitleCase();
        return eval("function " + functionName + "(next)\n" +
            "{\n" +
            "    performStandardTestCase('" + expression + "', next);\n" +
            "}; " + functionName);
    }

    InspectorTest.runTestSuite(expressions.map(createTestSuiteFunction));
}

</script>
</head>

<body onload="runTest()">
<p>
Tests that "Show Generator Location" jumps to the correct location.
</p>
</body>
</html>
