third_party/WebKit/LayoutTests/inspector/sources/debugger/show-generator-location-expected.txt

Tests that "Show Generator Location" jumps to the correct location.


Running: testIterNotStarted
Generator location revealed: [14:14]

Running: testIterSuspended1
Generator location revealed: [16:5]

Running: testIterSuspended2
Generator location revealed: [17:5]

Running: testIterSuspended3
Generator location revealed: [18:5]

Running: testIterClosed
Generator location revealed: [14:14]
